Add first processing function for standings data
Since the getStandings return data is quite complex, I want to create two major functions to process them. One to get all matchdata in a easy to read format, the other to get the standings in an easy to read format. This becomes big, so it was moved to a seperate file.